feat(regex): add literal trigram search
