+++ crates/core/src/regex_search/trigram_index.rs
@@ -1,66 +1,62 @@
 use std::{
     collections::{BTreeSet, HashMap},
-    fs,
-    path::{Path, PathBuf},
+    ops::Range,
+    path::Path,
 };
 
-use walkdir::WalkDir;
-
+use super::{
+    FileSystemCorpus, LiteralSearchResult, RegexCorpus, RegexSearchMatch, Trigram,
+    line_range_for_match,
+};
 use crate::index::{
     DocId,
     document_registry::{DocumentCatalog, DocumentMetadata, DocumentRegistry},
 };
 
-#[derive(Debug, Clone, PartialEq, Eq, Hash, PartialOrd, Ord)]
-pub struct Trigram(String);
-
-impl Trigram {
-    pub fn as_str(&self) -> &str {
-        &self.0
-    }
-}
-
-impl From<&str> for Trigram {
-    fn from(value: &str) -> Self {
-        Self(value.to_string())
-    }
-}
-
 #[derive(Debug)]
-pub struct TrigramIndex {
+pub struct TrigramIndex<C = FileSystemCorpus> {
+    corpus: C,
     postings: HashMap<Trigram, BTreeSet<DocId>>,
     documents: DocumentRegistry,
     doc_ids_by_path: Vec<DocId>,
 }
 
-impl TrigramIndex {
+impl TrigramIndex<FileSystemCorpus> {
     pub fn new(root: impl AsRef<Path>) -> Self {
+        Self::with_corpus(FileSystemCorpus::new(root.as_ref()))
+    }
+}
+
+impl<C> TrigramIndex<C>
+where
+    C: RegexCorpus,
+{
+    pub fn with_corpus(corpus: C) -> Self {
         let mut documents = DocumentRegistry::new();
         let mut postings: HashMap<Trigram, BTreeSet<DocId>> = HashMap::new();
-        let mut indexed_files = candidate_files(root.as_ref());
+        let mut corpus_documents = corpus.documents();
+        corpus_documents.sort_by(|left, right| left.path.cmp(&right.path));
 
-        indexed_files.sort();
-
-        for path in &indexed_files {
-            let Ok(content) = fs::read_to_string(path) else {
-                continue;
-            };
-
+        for document in &corpus_documents {
             let doc_id = documents.insert_or_update(
-                path.to_path_buf(),
-                trigram_count(&content),
-                content.len() as u64,
+                document.path.clone(),
+                trigram_count(&document.content),
+                document.content.len() as u64,
             );
 
-            for trigram in trigrams(&content) {
+            for trigram in trigrams(&document.content) {
                 postings.entry(trigram).or_default().insert(doc_id);
             }
         }
 
-        let mut doc_ids_by_path = documents_by_path(&documents, &indexed_files);
+        let mut doc_ids_by_path = corpus_documents
+            .iter()
+            .filter_map(|document| documents.doc_id(&document.path))
+            .collect::<Vec<_>>();
         doc_ids_by_path.sort();
 
         Self {
+            corpus,
             postings,
             documents,
             doc_ids_by_path,
@@ -103,6 +99,37 @@ impl TrigramIndex {
 
         candidates
     }
+
+    pub fn search_literal(&self, literal: &str) -> LiteralSearchResult {
+        if literal.is_empty() {
+            return LiteralSearchResult {
+                candidate_count: 0,
+                matches: Vec::new(),
+            };
+        }
+
+        let candidate_ids = self.candidates_for_literal(literal);
+        let candidate_count = candidate_ids.len();
+        let mut candidate_paths = candidate_ids
+            .iter()
+            .filter_map(|&doc_id| self.document(doc_id).map(|document| document.path.clone()))
+            .collect::<Vec<_>>();
+        candidate_paths.sort();
+
+        let mut matches = Vec::new();
+        for path in candidate_paths {
+            let Some(content) = self.corpus.read_document(&path) else {
+                continue;
+            };
+
+            matches.extend(verified_literal_matches(&path, &content, literal));
+        }
+
+        LiteralSearchResult {
+            candidate_count,
+            matches,
+        }
+    }
 }
 
 pub fn trigrams(content: &str) -> Vec<Trigram> {
@@ -117,112 +144,20 @@ fn trigram_count(content: &str) -> usize {
     content.chars().count().saturating_sub(2)
 }
 
-fn candidate_files(root: &Path) -> Vec<PathBuf> {
-    WalkDir::new(root)
-        .into_iter()
-        .filter_map(Result::ok)
-        .filter(|entry| entry.file_type().is_file())
-        .map(|entry| entry.path().to_path_buf())
+fn verified_literal_matches(path: &Path, content: &str, literal: &str) -> Vec<RegexSearchMatch> {
+    content
+        .match_indices(literal)
+        .map(|(start, matched_text)| {
+            literal_match(path, content, start..start + matched_text.len())
+        })
         .collect()
 }
 
-fn documents_by_path(documents: &DocumentRegistry, paths: &[PathBuf]) -> Vec<DocId> {
-    paths
-        .iter()
-        .filter_map(|path| documents.doc_id(path))
-        .collect()
-}
-
-#[cfg(test)]
-mod tests {
-    use super::{Trigram, TrigramIndex, trigrams};
-
-    #[test]
-    fn trigrams_are_overlapping_character_windows() {
-        let trigrams = trigrams("abcd");
-
-        let values = trigrams.iter().map(Trigram::as_str).collect::<Vec<_>>();
-        assert_eq!(values, ["abc", "bcd"]);
-    }
-
-    #[test]
-    fn trigrams_return_empty_for_short_strings() {
-        assert!(trigrams("").is_empty());
-        assert!(trigrams("a").is_empty());
-        assert!(trigrams("ab").is_empty());
-    }
-
-    #[test]
-    fn trigrams_include_punctuation_whitespace_and_preserve_case() {
-        let trigrams = trigrams("A b!");
-
-        let values = trigrams.iter().map(Trigram::as_str).collect::<Vec<_>>();
-        assert_eq!(values, ["A b", " b!"]);
-    }
-
-    #[test]
-    fn index_maps_every_indexed_doc_id_back_to_path() {
-        let temp = tempfile::tempdir().unwrap();
-        let first = temp.path().join("a.rs");
-        let second = temp.path().join("b.rs");
-        std::fs::write(&first, "abcdef").unwrap();
-        std::fs::write(&second, "uvwxyz").unwrap();
-
-        let index = TrigramIndex::new(temp.path());
-        let first_id = index.doc_id(&first).unwrap();
-        let second_id = index.doc_id(&second).unwrap();
-
-        assert_eq!(index.num_docs(), 2);
-        assert_eq!(index.document(first_id).unwrap().path, first);
-        assert_eq!(index.document(second_id).unwrap().path, second);
-    }
-
-    #[test]
-    fn postings_record_documents_containing_each_trigram() {
-        let temp = tempfile::tempdir().unwrap();
-        let first = temp.path().join("a.rs");
-        let second = temp.path().join("b.rs");
-        std::fs::write(&first, "abcdef").unwrap();
-        std::fs::write(&second, "zabczz").unwrap();
-
-        let index = TrigramIndex::new(temp.path());
-        let postings = index.postings(&Trigram::from("abc")).unwrap();
-
-        assert!(postings.contains(&index.doc_id(&first).unwrap()));
-        assert!(postings.contains(&index.doc_id(&second).unwrap()));
-    }
-
-    #[test]
-    fn literal_candidates_are_supersets_of_true_matches() {
-        let temp = tempfile::tempdir().unwrap();
-        let matching = temp.path().join("match.rs");
-        let false_positive = temp.path().join("false_positive.rs");
-        let missing = temp.path().join("missing.rs");
-        std::fs::write(&matching, "xxabcdefxx").unwrap();
-        std::fs::write(&false_positive, "abc bcd cde def").unwrap();
-        std::fs::write(&missing, "abxde").unwrap();
-
-        let index = TrigramIndex::new(temp.path());
-        let candidates = index.candidates_for_literal("abcdef");
-
-        assert!(candidates.contains(&index.doc_id(&matching).unwrap()));
-        assert!(candidates.contains(&index.doc_id(&false_positive).unwrap()));
-        assert!(!candidates.contains(&index.doc_id(&missing).unwrap()));
-    }
-
-    #[test]
-    fn short_literal_candidates_include_all_indexed_documents() {
-        let temp = tempfile::tempdir().unwrap();
-        let first = temp.path().join("a.rs");
-        let second = temp.path().join("b.rs");
-        std::fs::write(&first, "a").unwrap();
-        std::fs::write(&second, "b").unwrap();
-
-        let index = TrigramIndex::new(temp.path());
-        let candidates = index.candidates_for_literal("ab");
-
-        assert_eq!(candidates.len(), 2);
-        assert!(candidates.contains(&index.doc_id(&first).unwrap()));
-        assert!(candidates.contains(&index.doc_id(&second).unwrap()));
+fn literal_match(path: &Path, content: &str, byte_range: Range<usize>) -> RegexSearchMatch {
+    RegexSearchMatch {
+        path: path.to_path_buf(),
+        line_range: line_range_for_match(content, byte_range.clone()),
+        matched_text: content[byte_range.clone()].to_string(),
+        byte_range,
     }
 }
